// api-service-template/src/types.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */

export interface paths {
    "/sample-calc": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * 計算API
         * @description サンプルAPI、実行結果として与えられた数値の加算値を返却する。
         */
        get: {
            parameters: {
                query: {
                    /** @description パラメータ数値１ */
                    num1: string;
                    /** @description パラメータ数値２ */
                    num2: string;
                };
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description リクエスト成功時、返却オブジェクト */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["Result"];
                    };
                };
                /** @description エラー時、返却オブジェクト */
                400: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["Error"];
                    };
                };
                /** @description Unexpected error */
                default: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["Error"];
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
        Result: {
            /**
             * @description 実行日時
             * @example 2025/06/18 12:00:00
             */
            date?: string;
            /**
             * Format: int32
             * @description 計算結果
             * @example 123
             */
            "calc-result"?: number;
        };
        Error: {
            /**
             * @description エラーコード
             * @example 00001
             */
            code: string;
            /**
             * @description エラーメッセージ
             * @example 入力が数値型ではありません
             */
            message: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export type operations = Record<string, never>;